+++ examples/my_plugin.py
@@ -36,7 +36,8 @@ def mw2(request, context):
     print("Hello world")
 
 
-@my_plugin.middleware(priority=8, attach_to='response', relative='pre', with_context=True)
+@my_plugin.middleware(priority=8, attach_to='response', relative='pre',
+                      with_context=True)
 def mw3(request, response, context):
     assert 'test1' in context and context['test1'] == "testa"
     assert 'test2' in context and context['test2'] == "testb"
@@ -55,7 +56,8 @@ def t1(request):
     return text('from plugin!')
 
 def decorate(app, *args, **kwargs):
-    return my_plugin.decorate(app, *args, run_middleware=False, **kwargs)
+    return my_plugin.decorate(app, *args, with_context=True,
+                              run_middleware=True, **kwargs)
 
 
 __all__ = ["my_plugin", "decorate"]